+++ vector/src/main/java/im/vector/riotx/features/attachments/preview/AttachmentsPreviewAction.kt
@@ -17,10 +17,11 @@
 
 package im.vector.riotx.features.attachments.preview
 
+import android.net.Uri
 import im.vector.riotx.core.platform.VectorViewModelAction
 
 sealed class AttachmentsPreviewAction : VectorViewModelAction {
     object RemoveCurrentAttachment : AttachmentsPreviewAction()
     data class SetCurrentAttachment(val index: Int): AttachmentsPreviewAction()
-    data class UpdatePathOfCurrentAttachment(val newPath: String): AttachmentsPreviewAction()
+    data class UpdatePathOfCurrentAttachment(val newUri: Uri): AttachmentsPreviewAction()
 }