+++ CHANGES.md
@@ -3,12 +3,16 @@ Changes in RiotX 0.19.0 (2020-XX-XX)
 
 Features ✨:
  - Cross-Signing | Support SSSS secret sharing (#944)
+ - Cross-Signing | Verify new session from existing session (#1134)
 
 Improvements 🙌:
  - Verification DM / Handle concurrent .start after .ready (#794)
+ - Reimplementation of multiple attachment picker
 
 Bugfix 🐛:
  - Missing avatar/displayname after verification request message (#841)
+ - Crypto | RiotX sometimes rotate the current device keys (#1170)
+ - RiotX can't restore cross signing keys saved by web in SSSS (#1174)
 
 Translations 🗣:
  -